greenhouse/requirements.txt
confluent-kafka==2.11.1
cassandra-driver==3.29.2
